node: Add a NetworkUnavailable condition, for GCE
GCE complains about CIDRNotAvailable regularly if this condition isn't
ready. So stick it there to make GCE alias networking shut up.

This condition should probably be linked to the health of the Wireguard
tunnel. But the GCE controller won't be able to help out, so it would
really just be another way to say "Hey I am not ready right now".