Attempt at allowing tmpfs for EmptyDir
I figure this might be useful for e.g. Plex transcoding workdirs, where
pieces are written to persistent disk just to be delete (as the read
happens out of disk cache).

Not too sure the best way to implement this and still have the tmpfs's
memory usage accounted for as pod/container memory. I don't know if this
patch achieves that as-is. In fact I don't know how Kubernetes does it
if the EmptyDir is shared by multiple containers with different memory
profiles.

It'll be easier to see how this works if it's available, though.